requirements.txt
numpy
azure-functions